+++ Box2D/Box2D/Collision/Shapes/b2EdgeShape.h
@@ -49,7 +49,7 @@ public:
 	void ComputeAABB(b2AABB* aabb, const b2Transform& transform, int32_t childIndex) const;
 
 	/// @see b2Shape::ComputeMass
-	void ComputeMass(b2MassData* massData, float density) const;
+	void ComputeMass(b2MassData* massData, double density) const;
 	
 	/// These are the edge vertices
 	b2Vec2 m_vertex1, m_vertex2;
@@ -63,10 +63,10 @@ inline b2EdgeShape::b2EdgeShape()
 {
 	m_type = e_edge;
 	m_radius = b2_polygonRadius;
-	m_vertex0.x = 0.0f;
-	m_vertex0.y = 0.0f;
-	m_vertex3.x = 0.0f;
-	m_vertex3.y = 0.0f;
+	m_vertex0.x = 0.0;
+	m_vertex0.y = 0.0;
+	m_vertex3.x = 0.0;
+	m_vertex3.y = 0.0;
 	m_hasVertex0 = false;
 	m_hasVertex3 = false;
 }
