+++ Box2D/Box2D/Collision/Shapes/b2Shape.h
@@ -27,13 +27,13 @@
 struct b2MassData
 {
 	/// The mass of the shape, usually in kilograms.
-	float mass;
+	double mass;
 
 	/// The position of the shape's centroid relative to the shape's origin.
 	b2Vec2 center;
 
 	/// The rotational inertia of the shape about the local origin.
-	float I;
+	double I;
 };
 
 /// A shape is used for collision detection. You can create a shape however you like.
@@ -87,10 +87,10 @@ public:
 	/// The inertia tensor is computed about the local origin.
 	/// @param massData returns the mass data for this shape.
 	/// @param density the density in kilograms per meter squared.
-	virtual void ComputeMass(b2MassData* massData, float density) const = 0;
+	virtual void ComputeMass(b2MassData* massData, double density) const = 0;
 
 	Type m_type;
-	float m_radius;
+	double m_radius;
 };
 
 inline b2Shape::Type b2Shape::GetType() const
