+++ Box2D/Box2D/Collision/b2BroadPhase.h
@@ -97,7 +97,7 @@ public:
 	int32_t GetTreeBalance() const;
 
 	/// Get the quality metric of the embedded tree.
-	float GetTreeQuality() const;
+	double GetTreeQuality() const;
 
 	/// Shift the world origin. Useful for large worlds.
 	/// The shift formula is: position -= newOrigin
@@ -176,7 +176,7 @@ inline int32_t b2BroadPhase::GetTreeBalance() const
 	return m_tree.GetMaxBalance();
 }
 
-inline float b2BroadPhase::GetTreeQuality() const
+inline double b2BroadPhase::GetTreeQuality() const
 {
 	return m_tree.GetAreaRatio();
 }
