+++ Box2D/Box2D/Collision/b2Collision.h
@@ -69,8 +69,8 @@ union b2ContactID
 struct b2ManifoldPoint
 {
 	b2Vec2 localPoint;		///< usage depends on manifold type
-	float normalImpulse;	///< the non-penetration impulse
-	float tangentImpulse;	///< the friction impulse
+	double normalImpulse;	///< the non-penetration impulse
+	double tangentImpulse;	///< the friction impulse
 	b2ContactID id;			///< uniquely identifies a contact point between two shapes
 };
 
@@ -114,12 +114,12 @@ struct b2WorldManifold
 	/// point count, impulses, etc. The radii must come from the shapes
 	/// that generated the manifold.
 	void Initialize(const b2Manifold* manifold,
-					const b2Transform& xfA, float radiusA,
-					const b2Transform& xfB, float radiusB);
+					const b2Transform& xfA, double radiusA,
+					const b2Transform& xfB, double radiusB);
 
 	b2Vec2 normal;								///< world vector pointing from A to B
 	b2Vec2 points[b2_maxManifoldPoints];		///< world contact point (point of intersection)
-	float separations[b2_maxManifoldPoints];	///< a negative value indicates overlap, in meters
+	double separations[b2_maxManifoldPoints];	///< a negative value indicates overlap, in meters
 };
 
 /// This is used for determining the state of contact points.
@@ -147,7 +147,7 @@ struct b2ClipVertex
 struct b2RayCastInput
 {
 	b2Vec2 p1, p2;
-	float maxFraction;
+	double maxFraction;
 };
 
 /// Ray-cast output data. The ray hits at p1 + fraction * (p2 - p1), where p1 and p2
@@ -155,7 +155,7 @@ struct b2RayCastInput
 struct b2RayCastOutput
 {
 	b2Vec2 normal;
-	float fraction;
+	double fraction;
 };
 
 /// An axis aligned bounding box.
@@ -167,21 +167,21 @@ struct b2AABB
 	/// Get the center of the AABB.
 	b2Vec2 GetCenter() const
 	{
-		return 0.5f * (lowerBound + upperBound);
+		return 0.5 * (lowerBound + upperBound);
 	}
 
 	/// Get the extents of the AABB (half-widths).
 	b2Vec2 GetExtents() const
 	{
-		return 0.5f * (upperBound - lowerBound);
+		return 0.5 * (upperBound - lowerBound);
 	}
 
 	/// Get the perimeter length
-	float GetPerimeter() const
+	double GetPerimeter() const
 	{
-		float wx = upperBound.x - lowerBound.x;
-		float wy = upperBound.y - lowerBound.y;
-		return 2.0f * (wx + wy);
+		double wx = upperBound.x - lowerBound.x;
+		double wy = upperBound.y - lowerBound.y;
+		return 2.0 * (wx + wy);
 	}
 
 	/// Combine an AABB into this one.
@@ -242,7 +242,7 @@ void b2CollideEdgeAndPolygon(b2Manifold* manifold,
 
 /// Clipping for contact manifolds.
 int32_t b2ClipSegmentToLine(b2ClipVertex vOut[2], const b2ClipVertex vIn[2],
-							const b2Vec2& normal, float offset, int32_t vertexIndexA);
+							const b2Vec2& normal, double offset, int32_t vertexIndexA);
 
 /// Determine if two generic shapes overlap.
 bool b2TestOverlap(	const b2Shape* shapeA, int32_t indexA,
@@ -254,7 +254,7 @@ bool b2TestOverlap(	const b2Shape* shapeA, int32_t indexA,
 inline bool b2AABB::IsValid() const
 {
 	b2Vec2 d = upperBound - lowerBound;
-	bool valid = d.x >= 0.0f && d.y >= 0.0f;
+	bool valid = d.x >= 0.0 && d.y >= 0.0;
 	valid = valid && lowerBound.IsValid() && upperBound.IsValid();
 	return valid;
 }
@@ -265,10 +265,10 @@ inline bool b2TestOverlap(const b2AABB& a, const b2AABB& b)
 	d1 = b.lowerBound - a.upperBound;
 	d2 = a.lowerBound - b.upperBound;
 
-	if (d1.x > 0.0f || d1.y > 0.0f)
+	if (d1.x > 0.0 || d1.y > 0.0)
 		return false;
 
-	if (d2.x > 0.0f || d2.y > 0.0f)
+	if (d2.x > 0.0 || d2.y > 0.0)
 		return false;
 
 	return true;
